refactor: 불필요한 라인 제거
